A is for the strictly parametric terms
The definition of  $\mathbf{A_i}$ given in the manuscript is incorrect. **A** is the model matrix for the strictly parametric terms (i.e terms one might include in a LM or GLM). In general non-parametric is a bit of a misnomer, as all terms in the GAM as you (and Simon) define it are parametric once we've formed the basis expansion(s) for any smooth terms in the model. The distinction between what is and is not parametric gets complicated as models of this form:

```r
m <- gam(y ~ x1 * f1 + bs(x2) + s(x3) + te(x4, x5), data = df, method = "REML")
```

are perfectly valid but where the unpenalized B-spline (generated by the `bs()` call - I might need something to remove the intercept in that call) of `x2` would likely be part of **A** because the notation Simon uses for the smooths is restricted for the penalised smooths, which will have smoothing parameters associated with them.

But regardless, **A** is for the parametric terms.